Using None & Try,Except
A Program that repeatedly prompts a user for integer numbers until the user enters 'done'. Once 'done' is entered, prints out the largest and smallest of the numbers. If the user enters anything other than a valid number catches it with a try/except and puts out an appropriate message